requirements.txt
opencv-python
numpy<2.0
Pillow
ImageHash
pandas
easyocr
torch
torchvision
python-dotenv
requests